requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework-recursive==0.1.2
sqlparse==0.5.3
